TANGERINE-197: allow for free form entry for code lookups
git-svn-id: svn://svn.orangeleap.com/tangerine/mpower/trunk@3903 0f581344-5646-3a48-8ded-616bfea576c6

+++ src/com/mpower/domain/DistributionLine.java
@@ -44,7 +44,7 @@ public class DistributionLine implements Customizable, Viewable, Serializable {
     private BigDecimal amount;
     
     @Column(name = "PERCENTAGE")
-    private Short percentage;
+    private Float percentage;
 
     @Column(name = "PROJECT_CODE")
     private String projectCode;
@@ -116,11 +116,11 @@ public class DistributionLine implements Customizable, Viewable, Serializable {
         this.amount = amount;
     }
 
-    public Short getPercentage() {
+    public Float getPercentage() {
         return percentage;
     }
 
-    public void setPercentage(Short percentage) {
+    public void setPercentage(Float percentage) {
         this.percentage = percentage;
     }
